requirements.txt
discord.py
python-dotenv
mcstatus
pet-pet-gif
Pillow